Declare independence
Previously, Isolate was developed as a part of the Moe contest
system. Since most parts of Moe are not developed any longer, but
Isolate is used by several other contest systems, we decided to
fork it off.

We tried to preserve as much of Isolate's history as possible,
so all previous commits are extracted from Moe's Git repository
and stripped off unrelated parts.

Now, we replace Moe build system by a simple stand-alone Makefile
and a hand-edited config.h.

Also, we remove some relics.