Security: Avoid permission check bypass when bind-mounting directories
As the mount() syscall requires administrator privileges, we called
it as root. However, this could have been used to access paths whose
last component was accessible to the calling user, but some of the
previous components were not.

Now we call mount() with effective identity of the calling user with
the CAP_SYS_ADMIN capability added. This way, mount() is still possible,
but the file permissions are checked with respect to the calling user.

Please note that Isolate now requires the libcap library.