Makefile: Rewrote install target
"install -D" does not reliably create the directory if multiple
source files are given. This behavior of some versions of install
is not documented, so it is likely that it is a bug.

Changed the install target to create all directories explicitly.
It is a little bit verbose, but more straightforward.