By default, let stderr be inherited from the parent
WARNING: This is a change in behavior from the previous releases,
which defaulted to "stderr follows stdout".

However, the previous default made little sense and it caused
error messages to be lost in several real-life scenarios.

If you want to reproduce the previous behavior, use
"--stderr=/dev/fd/1". If you are not interested in error messages
at all, use "--stderr=/dev/null".