Introduced a proxy process
Previously, the program inside the sandbox was running as the init process
(PID 1) of its PID namespace. This interfered with signal handling as
there are special exceptions in delivering signals to init -- rougly
speaking, unhandled signals which would be otherwise fatal are ignored.

Now, we put a proxy process between the keeper and the inside process,
which serves as init and just passes the exit code to the keeper.