cg: add cg-oom-killed meta entry
It is hard to judge solely from the meta output whether a program has
been SIGKILL by the OOM killer or something else.

The cgroup API provides us with a memory.oom_control file that contains
a value "oom_kill" which contains the number of times the cgroup has
received an OOM kill. If this value is greater than 0, we report that
there was an OOM kill in the meta file with the cg-oom-killed entry.