Fixed time limits in non-cg mode
Introduction of the proxy process broke time limits in non-cg mode.
The box master is periodically polling the time spend by the child,
but the child is the proxy, not the inside process.

Now, we distinguish between box_pid and proxy_pid. Obtaining the box_pid
turned out to be tricky, but I hope it is correct.

Should fix #52, but needs testing.