Replaced the previous commit by a --tty-hack switch
I was resisting the temptation to add this mode for a long time,
because there are lots of non-obvious pitfalls when allowing untrusted
code to handle a tty.

However, there are some cases which are quite useful and reasonably
safe, like running a trusted text editor inside a sandbox. Still you
should be careful about which binaries can the user run from the editor.