Update makefile in order to actually work
The `config.h` file was removed a long time ago in favour of a dynamic config file. Unfortunately, it wasn't removed from the makefile and as such the %.o recipe fails to run (instead, it uses one of the make builtins). This simple commit makes the make recipe actually work.