Delete special files unless --special-files is given
Users of Isolate often forget to check if the output file they found
inside the sandbox is indeed a real file. Therefore they can be tricked
into opening a symlink (pointing to a file which was not accessible
inside the sandbox), or a named pipe (possibly deadlocking itself).
On some systems, you can also hard-link a device file from /dev to the
sandbox.

Since special files are seldom used as output, we decided to mitigate
the problem by removing all special files created inside the sandbox.
This happens during sandbox cleanup at the end of "isolate --run".

If you need special files to survive, please use the new --special-files
option.